health_wellness_agent/pyproject.toml
[project]
name = "health_wellness_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
